Use the interpreter's actual NUMERIC DIGITS rule
The rule was implemented as a fixed cap at 999999999. It is not a cap. The
C++ is `requestUnsignedNumber(setting, number_digits())`
(`NumericInstruction.cpp:103` for DIGITS, `:139` for FUZZ): the candidate is
rounded at the DIGITS *currently in force*, whatever decimals remain must
reduce to zero, and the result must not exceed `10^min(D,18) - 1`.

So the setting that is legal depends on the setting already in force:

    from DIGITS 3     numeric digits 1000                  error 26.005
    from DIGITS 1     numeric digits 10                    error 26.005
    from DIGITS 10    numeric digits 1000000000            ok
    from DIGITS 18    numeric digits 999999999999999999    ok

and rounding is part of it, which the first version of this fix missed
entirely because it only ever tried integers:

    from DIGITS 4     numeric digits 999.9999    ok, sets 1000
    from DIGITS 9     numeric digits 0.99999999995   ok, sets 1
    from DIGITS 3     numeric digits 999.6       error -- the carry lands
                                                 one position too wide
    from DIGITS 3     numeric digits 99.6        error -- the decimal does
                                                 not reduce to zero

That also rules out deciding this from a formatted rendering: at DIGITS 9,
`0.99999999995` renders as "1.00000000", which contains a decimal point in a
case the interpreter accepts. `unsignedNumberValue`, `checkIntegerDigits` and
`createUnsignedValue` are ported digit for digit instead, and pinned by a
1,056-case differential comparing outcome, sub-code and substitution values.

A bad FUZZ value now fails conversion as 26.006 rather than reaching the
fuzz-versus-digits comparison and reporting 33.001 -- the interpreter never
gets that far either.

Settings widen to `u64`, since 999999999999999999 is legal and `u32` cannot
hold it. That ripples through every operation's `digits` parameter; it is
carried rather than truncated at a boundary, because truncating would
reintroduce the defect below.

Three narrowing casts fixed with it. `muldiv.rs` divided through `digits as
i32`, the third instance of a class already fixed in `lib.rs` and `pow.rs`
and missed by two previous sweeps. The widening then made `2 * expt` overflow
at the low-end display trigger, found by the implementer in its own change and
fixed with `saturating_mul`.

`NUMERIC FORM VALUE` is case-sensitive: `'engineering'`, `'Engineering'`,
`'ENG'` and `' ENGINEERING'` are all rejected. Only the keyword written in
source is case-insensitive, and only because the tokenizer uppercases it
first. The comment claiming the interpreter "accepts any unambiguous case"
was false and is gone.

The old test passed because a fixed 999999999 cap and the real rule agree at
DIGITS 9 by coincidence -- nine digits versus ten. Every new settings test
starts from a non-default DIGITS.

Found by the whole-branch review. 170 tests pass, clippy clean, and all
eleven differential sets stay at 0 across 126,048 cases.

+++ rust/crates/rexx-num/src/compare.rs
@@ -82,7 +82,7 @@ impl CompareOp {
 /// `digits` and `fuzz` are `NUMERIC DIGITS`/`NUMERIC FUZZ`; `fuzz` is ignored
 /// for the strict operators, exactly as `NumberString::strictComp` takes no
 /// fuzz parameter at all.
-pub fn compare(a: &str, b: &str, digits: u32, fuzz: u32, op: CompareOp) -> Result<bool, ArithError> {
+pub fn compare(a: &str, b: &str, digits: u64, fuzz: u64, op: CompareOp) -> Result<bool, ArithError> {
     if op.is_strict() {
         // `primitiveIsEqual`/`primitiveStrictComp` are both a plain
         // shorter-prefix-then-length compare, with no blank stripping in
@@ -128,7 +128,7 @@ pub fn compare(a: &str, b: &str, digits: u32, fuzz: u32, op: CompareOp) -> Resul
 /// adding their magnitudes together (what a same-sign subtraction of them
 /// would do) could. The interpreter never attempts that computation for
 /// opposite signs; this must not either.
-fn numeric_order(a: &Number, b: &Number, digits: u32, fuzz: u32) -> Result<Ordering, ArithError> {
+fn numeric_order(a: &Number, b: &Number, digits: u64, fuzz: u64) -> Result<Ordering, ArithError> {
     let sign = |n: &Number| if n.is_zero() { 0 } else if n.negative { -1 } else { 1 };
     let (sign_a, sign_b) = (sign(a), sign(b));
     if sign_a != sign_b {
